10-blur_portion.c 2

+++ multithreading/10-blur_portion.c
@@ -10,9 +10,9 @@
  */
 pixel_t *apply_kernel(img_t *img, kernel_t *kernel, size_t px, size_t py)
 {
-size_t half = kernel->size / 2;
+int half = kernel->size / 2;
 float sum_r = 0, sum_g = 0, sum_b = 0, sum = 0;
-int kx, ky, nx,ny;
+int kx, ky, nx, ny;
 size_t pixel_idx;
 float kernel_value;
 pixel_t result;
